new/requirements-dev.txt
pytest-sugar
pytest-flake8
pytest-cov
pytest-mock
pytest-aiohttp
asynctest
